Moved web E2E to Test Suite
