+++ .circleci/config.yml
@@ -285,9 +285,9 @@ workflows:
   sdk:
     jobs:
       - expo_sdk
-      - web_test_suite
       # E2E SDK testing
-      - iOS-Detox
+      - web-Test-Suite
+      - iOS-Test-Suite
 
   # Android and iOS clients
   client:
@@ -368,7 +368,7 @@ jobs:
           name: Check packages
           command: expotools check-packages
 
-  web_test_suite:
+  web-Test-Suite:
     executor: web
     steps:
       - install_nix
@@ -380,8 +380,8 @@ jobs:
           working_directory: ~/expo
       - save_yarn_cache
       - yarn:
-          command: test:web:ci
-          working_directory: ~/expo/apps/test-suite
+          command: test:web
+          working_directory: ~/expo/apps/bare-expo
 
   home:
     executor: js
@@ -415,7 +415,7 @@ jobs:
           command: tsc
           working_directory: ~/expo/tools/expotools
 
-  iOS-Detox:
+  iOS-Test-Suite:
     executor: mac
     macos: # Detox simulators require 10.1.0
       xcode: '10.1.0'
