+++ apps/bare-expo/jest-puppeteer.config.js
@@ -1,0 +1,8 @@
+// For testing
+const { withExpoPuppeteer } = require('jest-expo-puppeteer');
+
+module.exports = withExpoPuppeteer({
+  launch: {
+    //   headless: false,
+  },
+});
