+++ packages/jest-expo-puppeteer/jest-puppeteer.config.js
@@ -3,5 +3,4 @@ const { withExpoPuppeteer } = require('./build');
 
 module.exports = withExpoPuppeteer({
   projectRoot: `tests/basic`,
-  mode: process.env.E2E_ENV,
 });
